feat: 3-1 코드 가져오기
